Android 5.1.0 release 1

+++ debuggerd/debuggerd.cpp
@@ -1,6 +1,5 @@
 /*
  * Copyright 2006, The Android Open Source Project
- * Copyright (c) 2013, The Linux Foundation. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -78,7 +77,7 @@ static void wait_for_user_action(const debugger_request_t &request) {
         "*\n"
         "* Wait for gdb to start, then press the VOLUME DOWN key\n"
         "* to let the process continue crashing.\n"
-        "********************************************************\n",
+        "********************************************************",
         request.pid, exe, request.tid);
 
   // Wait for VOLUME DOWN.
@@ -125,94 +124,16 @@ static int get_process_info(pid_t tid, pid_t* out_pid, uid_t* out_uid, uid_t* ou
   return fields == 7 ? 0 : -1;
 }
 
-static bool copy_file(const char* src, char* dest)
-{
-   #define BUF_SIZE 64
-   ssize_t bytes;
-   int  source_fh, dest_fh;
-   int  total_size = 0;
-   char buffer[BUF_SIZE];
-
-   if ((source_fh = open(src, O_RDONLY, O_NOFOLLOW)) == -1) {
-        ALOGE("Unable to open source file %s\n", src);
-   } else {
-        if((dest_fh = open(dest, O_WRONLY | O_CREAT | O_TRUNC | O_NOFOLLOW, 0640)) == -1) {
-            ALOGE("Unable to write destination file %s\n", dest);
-        } else {
-            while ((bytes = read(source_fh, buffer, BUF_SIZE)) > 0) {
-                if (write(dest_fh, buffer, bytes) < 0) {
-                    ALOGE("Write failed for destination file %s. Copied %d bytes\n",
-                            dest, total_size);
-                    break;
-                }
-                total_size += bytes;
-            }
-            ALOGI("Copied %s to %s - size: %d\n", src, dest, total_size);
-            fsync(dest_fh);
-            close(dest_fh);
-        }
-        close(source_fh);
-        if (total_size > 0) {
-            return true;
-        }
-    }
-    return false;
-}
-
-static void collect_etb_map(int cr_pid)
-{
-    struct stat s;
-    char   src_buf[64];
-    char   dest_buf[64];
-
-    snprintf(dest_buf, sizeof dest_buf, "/data/core/etb.%d", cr_pid);
-    if (!copy_file("/dev/coresight-tmc-etf", dest_buf)) {
-        ALOGE("Unable to copy ETB buffer file /dev/coresight-tmc-etf\n");
-    } else {
-        memset(src_buf, 0, sizeof(src_buf));
-        snprintf(src_buf, sizeof(src_buf), "/proc/%d/maps", cr_pid);
-        if(stat(src_buf, &s)) {
-            ALOGE("map file /proc/%d/maps does not exist for pid %d\n",
-                cr_pid, cr_pid);
-        } else {
-            snprintf(dest_buf, sizeof dest_buf, "/data/core/maps.%d", cr_pid);
-            if (!copy_file(src_buf, dest_buf)) {
-                ALOGE("Unable to copy map file /proc/%d/maps", cr_pid);
-            }
-        }
-    }
-}
-
-static void enable_etb_trace(struct ucred cr) {
-    char value[PROPERTY_VALUE_MAX];
-    property_get("persist.debug.trace", value, "");
-    if ((strcmp(value,"1") == 0)) {
-        /* Allow ETB collection only once; Note: in future this behavior can be changed
-        * To allow this, use a property to indicate whether the ETB has been collected */
-        property_get("debug.etb.collected", value, "");
-        if(strcmp(value,"1")) {
-            ALOGI("Collecting ETB dumps (from pid=%d uid=%d)\n",
-                     cr.pid, cr.uid);
-            property_set("debug.etb.collected", "1");
-            collect_etb_map(cr.pid);
-        }
-        else {
-            ALOGI("ETB already collected once, skipping (from pid=%d uid=%d)\n",
-                     cr.pid, cr.uid);
-        }
-    }
-}
-
 static int read_request(int fd, debugger_request_t* out_request) {
   ucred cr;
   socklen_t len = sizeof(cr);
   int status = getsockopt(fd, SOL_SOCKET, SO_PEERCRED, &cr, &len);
   if (status != 0) {
-    ALOGE("cannot get credentials\n");
+    ALOGE("cannot get credentials");
     return -1;
   }
 
-  ALOGV("reading tid\n");
+  ALOGV("reading tid");
   fcntl(fd, F_SETFL, O_NONBLOCK);
 
   pollfd pollfds[1];
@@ -249,7 +170,6 @@ static int read_request(int fd, debugger_request_t* out_request) {
     // Ensure that the tid reported by the crashing process is valid.
     char buf[64];
     struct stat s;
-    enable_etb_trace(cr);
     snprintf(buf, sizeof buf, "/proc/%d/task/%d", out_request->pid, out_request->tid);
     if (stat(buf, &s)) {
       ALOGE("tid %d does not exist in pid %d. ignoring debug request\n",
@@ -307,6 +227,7 @@ static void handle_request(int fd) {
       ALOGE("ptrace attach failed: %s\n", strerror(errno));
     } else {
       bool detach_failed = false;
+      bool tid_unresponsive = false;
       bool attach_gdb = should_attach_gdb(&request);
       if (TEMP_FAILURE_RETRY(write(fd, "\0", 1)) != 1) {
         ALOGE("failed responding to client: %s\n", strerror(errno));
@@ -320,8 +241,9 @@ static void handle_request(int fd) {
 
         int total_sleep_time_usec = 0;
         for (;;) {
-          int signal = wait_for_signal(request.tid, &total_sleep_time_usec);
-          if (signal < 0) {
+          int signal = wait_for_sigstop(request.tid, &total_sleep_time_usec, &detach_failed);
+          if (signal == -1) {
+            tid_unresponsive = true;
             break;
           }
 
@@ -388,27 +310,21 @@ static void handle_request(int fd) {
         free(tombstone_path);
       }
 
-      ALOGV("detaching\n");
-      if (attach_gdb) {
-        // stop the process so we can debug
-        kill(request.pid, SIGSTOP);
-
-        // detach so we can attach gdbserver
-        if (ptrace(PTRACE_DETACH, request.tid, 0, 0)) {
-          ALOGE("ptrace detach from %d failed: %s\n", request.tid, strerror(errno));
-          detach_failed = true;
+      if (!tid_unresponsive) {
+        ALOGV("detaching");
+        if (attach_gdb) {
+          // stop the process so we can debug
+          kill(request.pid, SIGSTOP);
         }
-
-        // if debug.db.uid is set, its value indicates if we should wait
-        // for user action for the crashing process.
-        // in this case, we log a message and turn the debug LED on
-        // waiting for a gdb connection (for instance)
-        wait_for_user_action(request);
-      } else {
-        // just detach
         if (ptrace(PTRACE_DETACH, request.tid, 0, 0)) {
-          ALOGE("ptrace detach from %d failed: %s\n", request.tid, strerror(errno));
+          ALOGE("ptrace detach from %d failed: %s", request.tid, strerror(errno));
           detach_failed = true;
+        } else if (attach_gdb) {
+          // if debug.db.uid is set, its value indicates if we should wait
+          // for user action for the crashing process.
+          // in this case, we log a message and turn the debug LED on
+          // waiting for a gdb connection (for instance)
+          wait_for_user_action(request);
         }
       }
